Moved SdlAppStatusListener into SdlApplication as SdlApplication.StatusListener. Is now package private.
Moved SdlLifecycleListener into SdlApplication as SdlApplication.LifecycleListener. Is now package private.

Created interfaces sub package to hold public facing interfaces like SdlContext.

Created SdlContextAbsImpl class as an abstract implementation of the SdlContext interface to define some common operations.
Unable to inherit from this class for SdlService in the current architecture. This needs to be revised or the interface will need to be implemented in the SdlService class.