Packets are prioritized based on waiting time and size.
Each app contains a serial queue of packets. The router service will iterate through each app peeking at the next packet in each of their respective queues. The service then sends the packet that has the highest priority and restarts this process. If no packets need to be sent it waits until one enters a queue.