SdlPackets now have a priority coefficient for the router service to determine priority.
This change touched a lot of classes as it needed to get to the multiplexing transport. Instead of sending a byte[] with offset and length, the entire SdlPacket is now passed through different layers until it reaches the SdlTransport. The router service uses this to lower priority based on if this is a multiframe message or a stream of rpcs.