fix: MultiplexTcpTransport may open socket via mobile network
On Android 5+, a Socket will be bound to mobile network if
the phone is connected to a Wi-Fi network with no Internet
connectivity. This "feature" prevents us from making a local
TCP connection between Core.
This fix makes sure that the TCP transport works with head
units that provide no Internet connectivity over their Wi-Fi.