Handle corner case during transport d/c
There is an issue in the first gen of multi transport router services that will remove certain extras from messages to the TransportBroker if older apps are connected that do not support the multi transport messages. Because of /that, we check the records we have and if the transport matches we assume it was the original transport that was received regardless of the address.