[SDL 0179]Pixel density and Scale
Fixes https://github.com/smartdevicelink/sdl_java_suite/issues/804

The main change is related to changing the size of the view captured.
This is done by adding a new scale parameter in the video
streaming capability and by dividing the preferred resolution from the
HMI and the scale.

Changes in the calculation of the touch coordinates are not needed due
to the array VideoStreamManager.touchScalar, this array is already
handling the differences in the size between the preferred resolution
and the captured view.
The touch coordinates are correctly being divided by the scale value.